# crates/genomehubs-api/target/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GenomeHubs API",
    "description": "GenomeHubs Application Programming Interface.\n\nProvides a low-level API to retrieve records, look up taxa, search and generate reports across all taxon, assembly, and sample data.\n\nSee the [GenomeHubs documentation](https://genomehubs.org) for further details.",
    "contact": {
      "name": "GenomeHubs",
      "url": "https://genomehubs.org"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/genomehubs/genomehubs/blob/main/LICENSE"
    },
    "version": "3.0.0"
  },
  "paths": {
    "/api/v3/count": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Count records matching a search query",
        "description": "Fetch the record count for a search query without including any of the results.\n\n`/count` supports all relevant `/search` parameters to allow a record count to be obtained for a query prior to a full search.",
        "operationId": "post_count",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountRequest"
              },
              "examples": {
                "Assembly count": {
                  "summary": "Count assemblies for Mammalia",
                  "value": {
                    "params_yaml": "size: 0\ntaxonomy: ncbi\n",
                    "query_yaml": "index: assembly\ntaxa:\n  - Mammalia\ntaxon_filter_type: tree\n"
                  }
                },
                "Mammalia species count": {
                  "summary": "Count species in Mammalia with a genome size estimate",
                  "value": {
                    "params_yaml": "size: 0\ninclude_estimates: true\ntaxonomy: ncbi\n",
                    "query_yaml": "index: taxon\ntaxa:\n  - Mammalia\ntaxon_filter_type: tree\nfields:\n  - name: genome_size\n"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Count result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/count/batch": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Count records for multiple queries in a single request",
        "description": "Execute up to 100 count queries simultaneously using a single POST. Reduces network round-trips for parallel counts.",
        "operationId": "post_count_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountBatchRequest"
              },
              "examples": {
                "Two counts": {
                  "summary": "Count Mammalia and Insecta taxa in parallel",
                  "value": {
                    "queries": [
                      {
                        "params_yaml": "taxonomy: ncbi\n",
                        "query_yaml": "index: taxon\nquery: tax_tree(Mammalia)\n"
                      },
                      {
                        "params_yaml": "taxonomy: ncbi\n",
                        "query_yaml": "index: taxon\nquery: tax_tree(Insecta)\n"
                      }
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch count results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountBatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/lookup": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Find record IDs matching a search term",
        "description": "Find record IDs matching a search term based on full or partial matches to a record ID, name, synonym or other identifier.",
        "operationId": "get_lookup",
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "description": "Search term for lookup",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "result",
            "in": "query",
            "description": "Result type (taxon|assembly|sample)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Maximum results to return (default 10)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lookup results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/lookup/batch": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Resolve up to 100 search terms to record IDs in one request",
        "description": "Runs each item through the same three-stage waterfall (SAYT → wildcard → suggest) as `GET /lookup`, using at most 3 `_msearch` calls for the entire batch. Results are returned in input order.",
        "operationId": "post_lookup_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch lookup results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupBatchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v3/metadata/fields": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "Returns available fields for a given result type",
        "description": "Returns field definitions, types, and synonyms for the specified result index (taxon, assembly, or sample).",
        "operationId": "get_result_fields",
        "parameters": [
          {
            "name": "result",
            "in": "query",
            "description": "Result type (taxon|assembly|sample)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Field metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultFieldsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v3/phylopic": {
      "get": {
        "tags": [
          "External"
        ],
        "summary": "Get PhyloPic silhouette for a taxon",
        "description": "Fetches a PhyloPic silhouette image record for a given NCBI taxon ID. Results are cached per API build cycle.",
        "operationId": "get_phylopic",
        "parameters": [
          {
            "name": "taxon_id",
            "in": "query",
            "description": "NCBI taxon ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taxonomy",
            "in": "query",
            "description": "Taxonomy name (e.g. ncbi)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PhyloPic silhouette record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhylopicResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/phylopic/batch": {
      "post": {
        "tags": [
          "External"
        ],
        "summary": "Resolve PhyloPic silhouettes for up to 200 taxa",
        "description": "Fetches PhyloPic silhouette image records for up to 200 NCBI taxon IDs in a single POST. Results are returned as a map keyed by taxon ID.",
        "operationId": "post_phylopic_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhylopicBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Per-taxon PhyloPic results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhylopicBatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/positional": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Generate a positional report (oxford / ribbon / painting)",
        "description": "Returns positional comparison data for one or more assemblies. Supports oxford (2-assembly dot-plot), ribbon (N-assembly synteny), and painting (single-assembly chromosome map).\n\n`query_yaml` is **optional**. When omitted, features are filtered by the `group_by` value (used as `primary_type`) automatically.\n\nSet `index: feature` in `query_yaml` to apply attribute filters directly to the feature index (e.g. `sequence_length >= 1000000`).\nSet `index: taxon` + `taxa:` to restrict features to a given taxon subtree.",
        "operationId": "post_positional",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionalRequest"
              },
              "examples": {
                "Oxford — feature attribute filter": {
                  "summary": "Oxford dot-plot filtered to large scaffolds only",
                  "value": {
                    "positional_yaml": "report: oxford\ngroup_by: busco_gene\nassemblies:\n  - GCA_905147045.1\n  - GCA_902806685.2\n",
                    "query_yaml": "index: feature\nattributes:\n  - name: sequence_length\n    operator: ge\n    value: \"1000000\"\n"
                  }
                },
                "Oxford — no query filter": {
                  "summary": "Oxford dot-plot, features filtered by group_by only",
                  "value": {
                    "positional_yaml": "report: oxford\ngroup_by: busco_gene\nassemblies:\n  - GCA_905147045.1\n  - GCA_902806685.2\nwindow_size: 1000000\n"
                  }
                },
                "Oxford — taxon filter": {
                  "summary": "Oxford dot-plot with taxon subtree filter",
                  "value": {
                    "positional_yaml": "report: oxford\ngroup_by: busco_gene\nassemblies:\n  - GCA_905147045.1\n  - GCA_902806685.2\nwindow_size: 1000000\n",
                    "query_yaml": "index: taxon\ntaxa:\n  - Lepidoptera\ntaxon_filter_type: tree\n"
                  }
                },
                "Painting": {
                  "summary": "Chromosome painting coloured by BUSCO status",
                  "value": {
                    "positional_yaml": "report: painting\ngroup_by: busco_gene\nassemblies:\n  - GCA_905147045.1\ncat: busco_status\nwindow_size: 500000\n"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Positional report data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/record": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Fetch a record by ID",
        "description": "Returns all field values for one or more specific records identified by their record ID.",
        "operationId": "get_record",
        "parameters": [
          {
            "name": "recordId",
            "in": "query",
            "description": "Record ID (comma-separated for multiple)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "result",
            "in": "query",
            "description": "Result type (taxon|assembly|sample)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record(s)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/record/batch": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Fetch up to 1,000 records by ID in one request",
        "description": "Accepts an explicit array of record IDs and returns all found records in a single `_mget` call. Prefer this over repeated `GET /record` calls.",
        "operationId": "post_record_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch record results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordBatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/report": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Generate an aggregated report",
        "description": "Returns aggregated data for reports including histograms, scatter plots, trees, and other visualisations.",
        "operationId": "post_report",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Report data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/search": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Fetch records matching a search query",
        "description": "Returns records matching a query. Supports all major query parameters including filtering, sorting, and pagination.",
        "operationId": "post_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              },
              "examples": {
                "Large mammal assemblies": {
                  "summary": "Search assemblies for mammals with genome >= 1 Gbp",
                  "value": {
                    "params_yaml": "size: 10\nfields: assembly_span,assembly_level\ntaxonomy: ncbi\n",
                    "query_yaml": "index: assembly\ntaxa:\n  - Mammalia\ntaxon_filter_type: tree\nfields:\n  - name: assembly_span\n    modifiers: [min=1000000000]\n"
                  }
                },
                "Mammalia with genome size": {
                  "summary": "Search for Mammalia taxa with a genome size estimate",
                  "value": {
                    "params_yaml": "size: 10\nfields: genome_size\ninclude_estimates: true\ntaxonomy: ncbi\n",
                    "query_yaml": "index: taxon\ntaxa:\n  - Mammalia\ntaxon_filter_type: tree\nfields:\n  - name: genome_size\n"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountBatchRequest": {
        "type": "object",
        "required": [
          "searches"
        ],
        "properties": {
          "searches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountBatchItem"
            },
            "description": "Array of queries to count in batch (max 100)."
          }
        }
      },
...
      "CountRequest": {
        "type": "object",
        "required": [
          "query_yaml",
          "params_yaml"
        ],
        "properties": {
          "params_yaml": {
            "type": "string"
          },
          "query_yaml": {
            "type": "string"
          }
        }
      },
      "CountResponse": {
        "type": "object",
        "required": [
          "status",
          "url"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "url": {
            "type": "string",
            "description": "The executed ES URL"
          }
        }
      },
...
      "LookupResponse": {
        "type": "object",
        "required": [
          "status",
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupResult"
            }
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        }
      },
...
    }
  },
...
}